add access control module in administrator | board session module | view agendas in sangguanian member module

+++ app/Http/Controllers/Admin/SanggunianMemberController.php
@@ -2,21 +2,26 @@
 
 namespace App\Http\Controllers\Admin;
 
+use Illuminate\Http\Request;
+use App\Services\UserService;
 use App\Models\SanggunianMember;
 use App\Pipes\User\ProfilePicture;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Pipeline;
-use App\Pipes\SanggunianMember\StoreSanggunianMember;
 use App\Services\SanggunianMemberService;
 use App\Repositories\SanggunianMemberRepository;
 use App\Http\Requests\SanggunianMemberStoreRequest;
 use App\Http\Requests\SanggunianMemberUpdateRequest;
+use App\Pipes\SanggunianMember\StoreSanggunianMember;
 use App\Pipes\SanggunianMember\UpdateSanggunianMember;
 
 final class SanggunianMemberController extends Controller
 {
+    private UserService $userService;
+
     public function __construct(private SanggunianMemberRepository $sanggunianMemberRepository, private SanggunianMemberService $sanggunianMemberService)
     {
+        $this->userService = app()->make(UserService::class);
     }
 
     public function index()
@@ -33,20 +38,10 @@ final class SanggunianMemberController extends Controller
 
     public function store(SanggunianMemberStoreRequest $request)
     {
-
         $pipe = Pipeline::send($request)->through([
             ProfilePicture::class,
             StoreSanggunianMember::class,
         ])->then(fn ($data) => $data);
-
-
-        // $this->sanggunianMemberRepository->store([
-        //     'fullname' => $request->fullname,
-        //     'district' => $request->district,
-        //     'sanggunian' => $request->sanggunian,
-        //     'username' => $request->username,
-        //     'password' => $request->password,
-        // ]);
 
         return back()->with('success', 'Successfully add new Sangguniang Panlalawigan Member');
     }
@@ -65,15 +60,16 @@ final class SanggunianMemberController extends Controller
                 ProfilePicture::class,
                 UpdateSanggunianMember::class,
             ])->then(fn ($data) => $data);
-        // $this->sanggunianMemberRepository->update($sanggunianMember, $this->sanggunianMemberService->isUserWantToChangePassword($request->all()));
 
         return back()->with('success', 'Success! Sangguniang Panlalawigan Member updated successfully.');
     }
 
-    public function destroy(SanggunianMember $sanggunianMember)
+    public function destroy(SanggunianMember $sanggunianMember, Request $request)
     {
-        $this->sanggunianMemberRepository->delete($sanggunianMember);
-
-        return back()->with('success', 'Successfully delete a record');
+        if ($this->userService->verify($request->key, auth()->user())) {
+            $this->sanggunianMemberRepository->delete($sanggunianMember);
+            return response()->json(['success' => true, 'message' => 'Record deleted successfully, this page will automatically refresh after 5 seconds to apply the changes.']);
+        }
+        return response()->json(['success' => false, 'message' => 'The credentials you provide is invalid.'], 422);
     }
 }
